Implemented product listing handler

+++ services/productsservice/internal/infrastructure/repository/product_model.go
@@ -7,7 +7,7 @@ type ProductModel struct {
 	ID          string               `gorm:"primaryKey;type:uuid;"`
 	ProductName string               `gorm:"uniqueIndex;not null"`
 	Description string               `gorm:"null"`
-	UnitPrice   int64                `gorm:"not null"`
+	UnitPrice   float64              `gorm:"not null"`
 	Stock       int64                `gorm:"not null"`
 	CategoryId  uint                 `gorm:"column:category_id;foreignKey:CategoryID;references:ID"`
 	Category    ProductCategoryModel `gorm:"foreignKey:CategoryID"`
